refactor: code split the filter options into their own component
